# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "api/du/training": {
            "post": {
                "tags": [
                    "Lucas Du"
                ],
                "summary": "Criar um novo treino",
                "description": "Cria um novo treino para um usuário.",
                "operationId": "2550d66227427d9d0d96de90404c115e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "weight",
                                    "repetitions",
                                    "time",
                                    "du_user_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Musculação"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "50kg"
                                    },
                                    "repetitions": {
                                        "type": "string",
                                        "example": "3x12"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "45 minutos"
                                    },
                                    "du_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Treino criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "training": {
                                            "$ref": "schemas/DuTraining"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Treino criado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "api/du/training/{id}": {
            "put": {
                "tags": [
                    "Lucas Du"
                ],
                "summary": "Atualizar um treino existente",
                "description": "Atualiza as informações de um treino existente.",
                "operationId": "dd87208c6babfced09d169eeb1e0ac51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do treino",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "weight",
                                    "repetitions",
                                    "time",
                                    "du_user_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Musculação"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "55kg"
                                    },
                                    "repetitions": {
                                        "type": "string",
                                        "example": "3x10"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "50 minutos"
                                    },
                                    "du_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Treino atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "training": {
                                            "$ref": "schemas/DuTraining"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Treino atualizado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Treino não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lucas Du"
                ],
                "summary": "Deletar um treino",
                "description": "Deleta um treino existente.",
                "operationId": "74c39dcfbcae9da9ce6d04bb5554a7ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do treino",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Treino deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Treino deletado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Treino não encontrado"
                    }
                }
            }
        },
...
    },
...
}